Added files Grid Search optimizations
Added 7 files DummyValveX,Y,Z, DummySensor,GridOptimize2D, OptUtils and OptPlot. Valves behave similarly to axis.
dummySensor is similar to dummyDAQ but a coordinate can be passed to it. It evaluates one of three functions at the given coordinate, Rosenbock,Parabol, or Ackley. GridOptimize2D does a scan of the coordinate space generating iterively smaller grids until an optimum gives a value that satisfies ftol or xtol. OptUtils and OptPlot are utility and plotting methods.